Add files via upload
Just testing to see if this is how you can actually use Github to collab. I used a template to try see if the pictures would even work for the background image design wise as well as help with just getting started. Definitely a beyond rough draft just to frame things.